requirements.txt
aiogram==3.4.1
python-dotenv==1.0.0
openai==1.12.0
fpdf==1.7.2
pytesseract==0.3.10
Pillow==10.3.0
aiohttp==3.9.5
